+++ src/main/java/Server/SSLServer.java
@@ -3,7 +3,7 @@ package Server;
 import Config.Config;
 import DataPack.DataPack;
 import Database.Database;
-import Room.Room;
+import GameObjects.Room;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
